Fix incorrect violation modal bug
When a new location is submitted ngRoute creates an entirely new view
(because the URL changes). This triggers a new Semantic UI modal as
well which has the same ID as the previous one. This modal from the
previous view gets shown on a category click and filters the old data.

The fix is to destroy the modal on the AngularJS view destroy step.
Closes #24.